# internal/web/static/openapi.yaml
openapi: 3.0.3
info:
  title: Knomit API
  version: 2.0.0
  description: |
    Git-backed knowledge base API (HAL/HATEOAS).

    All successful responses are `application/hal+json` (except SSE streams
    and the raw OpenAPI spec). Collection responses use an envelope with
    `count`, `_links`, and `_embedded.<rel>`. All error responses are
    RFC 9457 `application/problem+json`.
servers:
  - url: /api/v1
    description: v1
...
paths:
...
  /repos:
    get:
      tags: [repos]
      summary: List repositories
      operationId: listRepos
      responses:
        "200":
          description: Collection of repos
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/ReposCollection" }
        "500": { $ref: "#/components/responses/Problem" }
    post:
      tags: [repos]
      summary: Create a repository
      description: |
        Initiates repo creation and streams progress as newline-delimited JSON
        (NDJSON). Each line is a `CreateRepoEvent` object. The stream ends with
        a `done` or `error` event.
      operationId: createRepo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRepoRequest" }
      responses:
        "200":
          description: NDJSON stream of creation progress events
          content:
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/CreateRepoEvent"
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /repos/{repo}:
    parameters:
      - $ref: "#/components/parameters/Repo"
    get:
      tags: [repos]
      summary: Get repository
      operationId: getRepo
      responses:
        "200":
          description: Repository
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Repo" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [repos]
      summary: Archive a repository
      description: |
        Moves the repo to the archive. Returns the archived-repo record.
        Returns 409 if the repo is the default or the last remaining repo.
      operationId: archiveRepo
      responses:
        "200":
          description: Repository archived
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/ArchivedRepo" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches:
    parameters:
      - $ref: "#/components/parameters/Repo"
    get:
      tags: [branches]
      summary: List branches
      operationId: listBranches
      responses:
        "200":
          description: Collection of branches
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/BranchesCollection" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches/{branch}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
    get:
      tags: [branches]
      summary: Branch root with discoverable links
      operationId: getBranch
      responses:
        "200":
          description: Branch
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Branch" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /repos/{repo}/origin:
    parameters:
      - $ref: "#/components/parameters/Repo"
    get:
      tags: [origin]
      summary: Get configured remote origin
      operationId: getOrigin
      responses:
        "200":
          description: Origin configuration
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Origin" }
        "204": { description: No origin configured }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    put:
      tags: [origin]
      summary: Configure remote origin
      operationId: setOrigin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OriginConfigRequest" }
      responses:
        "200":
          description: Origin configured
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/OriginSetResponse" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [origin]
      summary: Remove origin configuration
      operationId: deleteOrigin
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/origin-sessions:
    parameters:
      - $ref: "#/components/parameters/Repo"
    get:
      tags: [origin]
      summary: List active origin sessions
      operationId: listOriginSessions
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OriginSessionSummary" }
        "404": { $ref: "#/components/responses/Problem" }
    post:
      tags: [origin]
      summary: Create origin session
      operationId: createOriginSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OriginConfigRequest" }
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [session_id]
                properties:
                  session_id: { type: string }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /archived:
    get:
      tags: [archive]
      summary: List archived repositories
      operationId: listArchivedRepos
      responses:
        "200":
          description: HAL collection of archived repos
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/ArchivedReposCollection" }
        "500": { $ref: "#/components/responses/Problem" }
  /archived/{id}/restore:
    parameters:
      - $ref: "#/components/parameters/ArchivedID"
    post:
      tags: [archive]
      summary: Restore an archived repository
      description: |
        Moves an archived repo back to active. Optionally renames it via
        `new_name`. Returns 409 if the name or origin is already in use by an
        active repo.
      operationId: restoreArchivedRepo
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RestoreRepoRequest" }
      responses:
        "200":
          description: Repository restored
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Repo" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /archived/{id}:
    parameters:
      - $ref: "#/components/parameters/ArchivedID"
    delete:
      tags: [archive]
      summary: Permanently purge an archived repository
      operationId: purgeArchivedRepo
      responses:
        "204": { description: Purged }
        "404": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches/{branch}/mcp:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
      - { name: profile, in: query, schema: { type: string, default: code } }
    get: &mcp
      tags: [mcp]
      summary: MCP server proxy
      description: |
        Dispatches to the configured MCP handler for the given profile.
        Accepts any HTTP method; request/response shape is defined by the
        Model Context Protocol.
      operationId: mcpDispatch
      responses:
        "200": { description: MCP response (shape varies) }
        "404": { $ref: "#/components/responses/Problem" }
    post: *mcp
    put: *mcp
    delete: *mcp
...
components:
  parameters:
    Repo:       { name: repo,      in: path, required: true, schema: { type: string } }
    Branch:     { name: branch,    in: path, required: true, schema: { type: string } }
...
    ArchivedID: { name: id,        in: path, required: true, schema: { type: string }, description: Archived-repo ID }
...
  responses:
    Problem:
      description: RFC 9457 problem+json
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Link:
      type: object
      required: [href]
      properties:
        href:      { type: string }
        templated: { type: boolean }
    LinkMap:
      type: object
      additionalProperties: { $ref: "#/components/schemas/Link" }
    CollectionBase:
      type: object
      required: [count, _links]
      properties:
        count:     { type: integer }
        _links:    { $ref: "#/components/schemas/LinkMap" }
        _embedded: { type: object }
    Problem:
      type: object
      required: [title, status]
      properties:
        type:     { type: string, default: "about:blank" }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }

    # Discovery / repos / branches
...
    RepoSummary:
      type: object
      required: [name, _links]
      properties:
        name:   { type: string }
        _links: { $ref: "#/components/schemas/LinkMap" }
    ReposCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                repos:
                  type: array
                  items: { $ref: "#/components/schemas/RepoSummary" }
    Repo:
      type: object
      required: [name, _links]
      properties:
        name:         { type: string }
        agent_branch: { type: string }
        description:
          type: string
          description: |
            Verbatim content of the repo's kb.md root manifest, read at HEAD.
            Omitted when the repo has no readable kb.md.
        _links: { $ref: "#/components/schemas/LinkMap" }
    BranchSummary:
      type: object
      required: [name, _links]
      properties:
        name:   { type: string }
        _links: { $ref: "#/components/schemas/LinkMap" }
    BranchesCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                branches:
                  type: array
                  items: { $ref: "#/components/schemas/BranchSummary" }
    Branch:
      type: object
      required: [name, head, _links]
      properties:
        name:               { type: string }
        head:               { type: string, description: HEAD commit hash }
        index_commit:       { type: string }
        embeddings_enabled: { type: boolean }
        is_agent_branch:    { type: boolean }
        _links:             { $ref: "#/components/schemas/LinkMap" }

    # Facts
...
    CreateRepoOrigin:
      type: object
      description: Remote origin to clone from (used when mode=clone).
      properties:
        url:         { type: string, description: Remote URL }
        branch:      { type: string, description: Remote branch to track }
        auth_method: { type: string, enum: [token, basic, ssh] }
        auth_token:  { type: string }
    CreateRepoRequest:
      type: object
      required: [name, mode]
      properties:
        name:
          type: string
          description: Unique repo name (slug).
        mode:
          type: string
          enum: [preset, custom, clone]
          description: |
            preset — use a built-in ontology preset (e.g. default, code).
            custom — supply ontology_yaml directly.
            clone  — clone from a remote origin; ontology is taken from the remote.
        ontology_preset:
          type: string
          description: "Preset identifier (e.g. default, code). Used when mode=preset."
        ontology_yaml:
          type: string
          description: Raw ontology YAML. Used when mode=custom.
        origin:
          $ref: "#/components/schemas/CreateRepoOrigin"
    CreateRepoEvent:
      type: object
      description: |
        One line of the NDJSON creation stream. Each object carries a `type`
        discriminator and optional progress fields. The stream always ends with
        a `done` or `error` event.
      required: [type]
      properties:
        type:
          type: string
          enum: [progress, done, error]
          description: Event kind.
        step:
          type: string
          description: Machine-readable step identifier (e.g. "init", "clone", "index").
        message:
          type: string
          description: Human-readable progress message.
        pct:
          type: number
          format: double
          description: Completion percentage (0–100), present on progress events.
        repo:
          type: string
          description: Repo name, present on the final done event.
        title:
          type: string
          description: Short summary, present on done/error events.
        detail:
          type: string
          description: Extended detail or error description.
    ArchivedRepo:
      type: object
      required: [id, name, archivedAt, _links]
      properties:
        id:
          type: string
          description: Stable identifier for the archived repo record.
        name:
          type: string
          description: Repo name at the time of archival.
        origin:
          type: string
          description: Remote origin URL, if configured.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the repo was archived.
        _links: { $ref: "#/components/schemas/LinkMap" }
    ArchivedReposCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                archived:
                  type: array
                  items: { $ref: "#/components/schemas/ArchivedRepo" }
    RestoreRepoRequest:
      type: object
      properties:
        new_name:
          type: string
          description: Optional new name for the restored repo (defaults to original name).

    # Origin
    Origin:
      type: object
      required: [url, _links]
      properties:
        name:        { type: string }
        url:         { type: string }
        branch:      { type: string }
        auth_method: { type: string, enum: [token, basic, ssh] }
        _links:      { $ref: "#/components/schemas/LinkMap" }
    OriginConfigRequest:
      type: object
      required: [url]
      properties:
        url:         { type: string }
        auth_method: { type: string, enum: [token, basic, ssh] }
        token:       { type: string }
        user:        { type: string }
        password:    { type: string }
    OriginSetResponse:
      type: object
      required: [status, _links]
      properties:
        status: { type: string, enum: [ok] }
        _links: { $ref: "#/components/schemas/LinkMap" }
    OriginSessionSummary:
      type: object
      required: [session_id, state]
      properties:
        session_id: { type: string }
        state:      { type: string, enum: [created, tested, previewed, applied, committed, error] }
        url:        { type: string }
...